+++ packages/core/src/utilities/index.ts
@@ -100,6 +100,7 @@ import calculateSpacingBetweenImageIds from './calculateSpacingBetweenImageIds';
 export * as logger from './logger';
 import { calculateNeighborhoodStats } from './calculateNeighborhoodStats';
 import getPixelSpacingInformation from './getPixelSpacingInformation';
+import { asArray } from './asArray';
 export { updateReferencedPlane } from './updateReferencedPlane';
 
 const getViewportModality = (viewport: IViewport, volumeId?: string) =>
@@ -204,4 +205,5 @@ export {
   buildMetadata,
   calculateNeighborhoodStats,
   getPixelSpacingInformation,
+  asArray,
 };