+++ packages/tools/src/utilities/contourSegmentation/index.ts
@@ -1,6 +1,7 @@
 import areSameSegment from './areSameSegment';
 import convertContourSegmentationAnnotation from './convertContourSegmentation';
-import { addition, subtraction } from './logicalOperators';
+import { copyAnnotation, copyContourSegment } from './copyAnnotation';
+export * from './logicalOperators';
 export { default as isContourSegmentationAnnotation } from './isContourSegmentationAnnotation';
 export { addContourSegmentationAnnotation } from './addContourSegmentationAnnotation';
 export { removeContourSegmentationAnnotation } from './removeContourSegmentationAnnotation';
@@ -10,17 +11,13 @@ export { contourSegmentationOperation } from './contourSegmentationOperation';
 export * from './sharedOperations';
 
 export {
-  addition,
-  subtraction,
   areSameSegment,
   convertContourSegmentationAnnotation,
+  copyContourSegment,
+  copyAnnotation,
 };
 
-export {
-  unifyPolylineSets,
-  unifyMultiplePolylineSets,
-  unifyAnnotationPolylines,
-  subtractPolylineSets,
-  subtractMultiplePolylineSets,
-  subtractAnnotationPolylines,
-} from './unifyPolylineSets';
+export * from './polylineUnify';
+export * from './polylineSubtract';
+export * from './polylineIntersect';
+export * from './polylineXor';